+++ README.md
@@ -27,6 +27,6 @@ containers or as virtual machines.
 | ingesters   | Split the Crypt4GH header and move the remainder to the storage backend. No cryptographic task, nor connection to the keyserver. |
 | verifiers   | Connect to the keyserver (via SSL) and decrypt the stored files and checksum them against their embedded checksum. |
 | vault       | Storage backend: as a regular file system or as a S3 object store. |
-| ID mapper   | Handles the so-called _Stable ID_ filename mappings from CentralEGA. |
+| Finalize    | Handles the so-called _Stable ID_ filename mappings from CentralEGA. |
 
 Find the [LocalEGA documentation](http://localega.readthedocs.io) hosted on [ReadTheDocs.org](https://readthedocs.org/).
