+++ deploy/tests/src/test/resources/config.properties
@@ -11,5 +11,5 @@ container.label.inbox = inbox
 container.label.ingest = ingest
 container.label.keys = keys
 container.label.mq = mq
-container.label.s3 = s3
+container.label.s3 = vault
 container.label.verify = verify
